requirements.txt
Authlib==1.1.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
ecdsa==0.18.0
Flask==2.0.1
Flask-Cors==3.0.10
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.0
google-cloud-core==2.3.2
google-cloud-datastore==2.9.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
libcst==0.4.7
MarkupSafe==2.1.1
mypy-extensions==0.4.3
proto-plus==1.22.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
python-dotenv==0.21.0
python-jose==3.3.0
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2